% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBhavcopy.R
\name{getBhavcopy}
\alias{getBhavcopy}
\title{getBhavcopy}
\usage{
getBhavcopy(startDate = as.Date("2010-01-01", order = "ymd"),
  endDate = as.Date("2010-01-10", order = "ymd"), fileName = NULL,
  weekends = FALSE, freq = TRUE)
}
\arguments{
\item{startDate}{- Date in %Y-%m-%d format.}

\item{endDate}{- Date in %Y-%m-%d format.}

\item{fileName}{- Name of the file where data to be stored with complete path.}

\item{weekends}{- default, it won't pull data for weekends. If needed enable the flag. Sometime Muhurat trade can be on weekends.}
}
\value{
Returns a dataframe with all the accumulated data.
}
\description{
This gets the bhavcopy from nse-india for given days. It will by default, do not download Sun/Sat.
}
\details{
trades - Returns the empty trades data frame with given 'n' objects
}
\note{
This will create a trade object
}
\author{
Siva Sunku
}
\keyword{bhavcopy}
