% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOneMinData.R
\name{getOneMinQuotes}
\alias{getOneMinQuotes}
\title{getOneMinQuotes}
\usage{
getOneMinQuotes(objectName, stock = NULL, fut = NULL, trim = "T23:55")
}
\arguments{
\item{objectName}{- objectname from which quotes to be read}

\item{stock}{- Stock name to be filtered. Default ALL stocks}

\item{fut}{- default NULL. If fut prices are needed F1/F2 to be passed. currently only F1/F2 are available}

\item{trim}{- If True, only prices till 15:30 hours are returned. Default NULL.}
}
\description{
getOneMinQuotes
}
\details{
This gets the quotes of a particular stock from the big one min data object. This returns the object as xts for that particular stock.
}
