% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStockQuotesNSE.R
\name{getStockQuotesNSE}
\alias{getStockQuotesNSE}
\title{getStockQuotesNSE}
\usage{
getStockQuotesNSE(stock = NULL, fileName = NULL, objectName = NULL)
}
\arguments{
\item{stock}{- Equity data to be downloaded}

\item{fileName}{- File name where the quotes(bhavcopy) are stored}

\item{objectName}{- if the data is stored as rda. Either filename or objectname should be there, not both and neither none.}

\item{startDate}{- Date in YMD format.}

\item{endDate}{- Date in YMD format.}
}
\value{
Returns a dataframe with all the OHLC data
}
\description{
This gets the daily data from the bhavcopy csv into a data frame(OHLC)
}
\details{
Copies the bhavcopies from nse and keeps the csv (yearlywise) in the given directory.
}
\note{
This will return an OHLC object with all the details for a given stock.
It will read the fileName(BHAVcopy archive), and return the data specific to that particular stock as a OHLC dataframe.
}
\author{
Siva Sunku
}
\keyword{getStockQuotes}
