% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOneMinData.R
\name{makeOneMinDataObj}
\alias{makeOneMinDataObj}
\title{makeOneMinDataObj}
\usage{
makeOneMinDataObj(dirName = ".", echoFlag = FALSE)
}
\arguments{
\item{dirName}{- input Directory name}

\item{echoFlag}{- flag to display the file being processed}
}
\description{
This makes the onemindata obtained from google drive into a data table
}
\details{
loads the data in a file(oneminute google drive data into a table). 
    It reads all the files in a directory.
       If file is a zip file, it needs to be processed ==== TBD === 
       Reads the text file and formats it into a data table with (Symbol, date, time , OHLC, volume)
       Returns the object with all the data read.
}
\author{
Siva Sunku
}
